+++ R/class_balanced.R
@@ -14,10 +14,11 @@ balanced_class <- R6::R6Class(
   ),
   public = list(
     report_start = function() {
+      action <- 
       private$.bar_envir <- new.env(parent = globalenv())
       private$.bar_id <- cli::cli_progress_bar(
         name = paste("pipeline", Sys.time()),
-        format = "{cli::pb_bar} {cli::pb_current}/{cli::pb_total}",
+        format = "processing targets {cli::pb_bar} {cli::pb_current}/{cli::pb_total}",
         .envir = .subset2(private, ".bar_envir"),
         auto_terminate = FALSE,
         clear = TRUE