+++ NEWS.md
@@ -1,4 +1,4 @@
-# targets 1.10.1.9010 (development)
+# targets 1.10.1.9011 (development)
 
 ## Changes to default settings
 
@@ -23,6 +23,7 @@
 * Create a new `"balanced"` reporter with a `cli` progress bar (#1442).
 * Deprecate reporters `"forecast"`, `"forecast_interactive"`, `"verbose_positives"`, and `"timestamp_positives"` (#1442).
 * Ensure colors printed to the console are preserved when forwarded from the `callr` process (#1442).
+* Add `tar_option_with()` (https://github.com/ropensci/tarchetypes/issues/215, @noamross).
 
 # targets 1.10.1